requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
django==4.1.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==3.13.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   django-rest-swagger
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
openapi-codec==1.3.2
    # via django-rest-swagger
pillow==9.3.0
    # via -r requirements.in
pytz==2022.6
    # via djangorestframework
requests==2.28.1
    # via coreapi
simplejson==3.17.6
    # via django-rest-swagger
sqlparse==0.4.3
    # via django
tzdata==2022.6
    # via django
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.12
    # via requests
